Bitpanda: fixes for unobtainable prices for ETHW and BEST
Background info: Bitpanda Pro, which is used to obtain historical prices for (crypto-)assets
is now ONE TRADING and a separate company to which Bitpanda only holds a minority stake.
After that split, BEST can't be used anymore on ONE TRADING and historical BEST price data isn't available anymore.
Also, ETHW, which stands for the Proof of Work branch of the ETH chain, isn't available on ONE TRADING.

This means, that we need to use the prices we have from the csv file as best as we can.
For normal transactions, that's possible by using the "Asset market price" column of the export but for fee withdrawals
from the BEST wallet, no price was actually associated with the withdrawal transaction.
For now, I'm using an asset price of 0 because I don't know how to fix this otherwise.

For the normal transactions, a new property has been added to the Operation class (`exported_price`) to be able to
carry the asset price for normal transactions into the processing classes like PriceData and use it there
if a proper price can't be obtained (see `get_cost` method in `PriceData` in price_data.py).

Source: https://support.bitpanda.com/hc/en-us/articles/9374684386332-Why-is-Bitpanda-Pro-evolving-to-become-One-Trading

* general: Added new data field `exported_price` to `Operation` class
* bitpanda general: For ETHW and BEST, the `exported_price` will be used because the ONE TRADING API doesn't return anything.
* bitpanda general fix: The `change` calculation used a change from an earlier data record for some operation types,
  because the `change` variable was not touched (for Airdrops of any kind and Staking operations of any kind).
  The missing operations are now added in `_read_bitpanda` in book.py and an Exception has been added for anything else not
  included in my change.